Consistently capitalize project components names
While not having a strong opinion on what the best choice can be here, let's
normalize to the most already used style which seems to be capitalizing both
Theo and the component name. CLI, being an acronym will be always fully
capitalized. All of this only applies to names used in sentences.